requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-dotenv
pydantic[email]
pyjwt[crypto]
passlib[bcrypt]
python-multipart
python-jose[cryptography]